requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.8.1
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831749120/work
chardet @ file:///C:/ci/chardet_1607690654534/work
charset-normalizer==2.0.10
conda==4.10.3
conda-package-handling @ file:///C:/ci/conda-package-handling_1618262320430/work
cryptography @ file:///C:/ci/cryptography_1616769344312/work
discord-buttons-plugin==0.1.2
discord-components==2.1.2
discord.py==1.7.3
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
menuinst==1.4.16
MouseInfo==0.1.3
multidict==5.2.0
mysql-connector-python==8.0.27
numpy==1.21.5
opencv-python==4.5.5.62
pandas==1.3.5
Pillow==8.4.0
protobuf==3.19.1
py-cord @ git+https://github.com/Pycord-Development/pycord@fc5556f808950c9757d2a1f3f817ec54ef9b4711
PyAutoGUI==0.9.53
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==227
requests==2.27.1
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016967756/work
schedule==1.1.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
yarl==1.7.2
